Change affine subspace standardisation, fixing issue #2.
The representative point of an affine subspace is standardised in
one of two ways, depending on the situation. It is either reduced
modulo lattice vectors, or reduced also modulo vectors parallel to
the space. Unfortunately, the latter reduction was not always reduced
according to the former. This commit makes the two reductions
compatible, as they should be.