Properly fix comparison of Left Wyckoff positions
The prior fix, checking if a single matrix was left-acting within ImageAffineSubspaceLattice, led to spurious results because that isn't a reliable check. Instead, we now ensure that the group U given to ImageAffineSubspaceLattice is right acting. This was done in other places it had been called, but not in the comparisons.