Partial fix: some cases with shifted origin work now
IsTranslationInBasis did not account for the possibility of the origin not being at zero. A more thorough check fixes some cases.
But not all cases. It's still possible for the solver to give a translation that is out of basis, but if a different translation mod 1 is chosen then it would be in basis.
I think the proper solution may require rebuilding all the modZ portions.