Implement WyPosSGL and WyckoffPositionsByStabilizer for subperiodic groups
This was as simple as substituting in the replacement functions and tests I had implemented for WyPosAT.

I spotted a potential bug in testing WyckoffPositionsByStabilizer, but that bug is also present when using a space group, so I'll leave it to someone else to fix.
Specifically, the routine will also return Wyckoff positions with a subgroup of the stabilizer if the positions share a basis.