WIP rewrite of SymmetricInternalBasis
Currently a regression, but necessary steps to improve mathematical robustness.
We do not want out basis to be strongly dependent on the order we read the operators. Consistency is important such that transforming to Standard basis actually is standard basis.
Current strategy involves grabbing all the potential basis vectors then Reducing them. Currently the results are hit-or-miss.